requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
starlette==0.46.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
